# [mono] Set t_currentThread in InitializeCurrentThread