@vargaz 2020-08-27
Why is this faster ?