@EgorBo 2020-08-27
@vargaz hm.. good question
I noticed a difference: CoreCLR inits a static field `t_currentThread` https://github.com/dotnet/runtime/blob/master/src/coreclr/src/System.Private.CoreLib/src/System/Threading/Thread.CoreCLR.cs#L275

Mono does not: https://github.com/dotnet/runtime/blob/master/src/mono/netcore/System.Private.CoreLib/src/System/Threading/Thread.Mono.cs#L324-L326 - it always icalls for get_CurrentThread and then we access a managed field.

So the speed improvement is probably not really worth it. (on corefx it's like 4x improvement)