@vargaz 2020-08-27
Maybe we should try to speed up CurrentThread instead.