@vargaz 2020-08-27
The CurrentThread implementation is very complicated on .net framework mono due to domains, but on netcore, it could just be a ThreadStatic or something.